Move the InstallFromImageTask class
It should be in the installation tasks of the live image payload module.
