+++ pyanaconda/modules/payloads/payload/live_image/installation.py
@@ -18,6 +18,7 @@
 import hashlib
 
 from pyanaconda.anaconda_loggers import get_module_logger
+from pyanaconda.core.constants import INSTALL_TREE
 from pyanaconda.core.i18n import _
 from pyanaconda.core.util import execWithRedirect, lowerASCII
 from pyanaconda.modules.common.task import Task
@@ -115,4 +116,53 @@ class InstallFromTarTask(Task):
             log.info(msg)
 
         if err:
+            raise PayloadInstallationError(err or msg)
+
+
+class InstallFromImageTask(Task):
+    """Task to install the payload from image."""
+
+    def __init__(self, dest_path, source=None):
+        """Create a new task.
+
+        :param dest_path: installation destination root path
+        :type dest_path: str
+        """
+        super().__init__()
+        self._source = source
+        self._dest_path = dest_path
+
+    @property
+    def name(self):
+        return "Install the payload from image"
+
+    def run(self):
+        """Run installation of the payload from image."""
+        # TODO: remove this check for None when Live Image payload will support sources
+        # The None check is just a temporary hack that Live OS has source but Live Image don't
+        if self._source is not None and not self._source.get_state():
+            raise PayloadInstallationError("Source is not set up!")
+
+        cmd = "rsync"
+        # preserve: permissions, owners, groups, ACL's, xattrs, times,
+        #           symlinks, hardlinks
+        # go recursively, include devices and special files, don't cross
+        # file system boundaries
+        # TODO: source will provide us source path instead of using constant here
+        args = ["-pogAXtlHrDx", "--exclude", "/dev/", "--exclude", "/proc/", "--exclude", "/tmp/*",
+                "--exclude", "/sys/", "--exclude", "/run/", "--exclude", "/boot/*rescue*",
+                "--exclude", "/boot/loader/", "--exclude", "/boot/efi/loader/",
+                "--exclude", "/etc/machine-id", INSTALL_TREE + "/", self._dest_path]
+        try:
+            rc = execWithRedirect(cmd, args)
+        except (OSError, RuntimeError) as e:
+            msg = None
+            err = str(e)
+            log.error(err)
+        else:
+            err = None
+            msg = "%s exited with code %d" % (cmd, rc)
+            log.info(msg)
+
+        if err or rc == 11:
             raise PayloadInstallationError(err or msg)
